Move dd class functions into their own file
