+++ src/adiar/zdd.h
@@ -20,6 +20,9 @@
 
 #include <adiar/bool_op.h>
 #include <adiar/file.h>
+
+#include <adiar/internal/dd_func.h>
+
 #include <adiar/zdd/zdd.h>
 #include <adiar/bdd/bdd.h>
